Add updated info from kafka listener

+++ src/main/java/com/inditex/zarachallenge/infrastructure/JpaProductRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JpaProductRepository extends JpaRepository<ProductEntity, Integer> {
+public interface JpaProductRepository extends JpaRepository<ProductEntity, Long> {
 }
